Windows: afslogon cast away warnings
cast away size_t to int warnings in 64-bit builds.

Reviewed-on: http://gerrit.openafs.org/7634
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Tested-by: Jeffrey Altman <jaltman@your-file-system.com>
(cherry picked from commit 76ae0f01b32d3e87273ff6d2c9c20eec59e22e16)

Reviewed-on: http://gerrit.openafs.org/7706
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ src/WINNT/afsd/afslogon.c
@@ -1329,11 +1329,11 @@ VOID AFS_Logoff_Event( PWLX_NOTIFICATION_INFO pInfo )
 	    size_t szlen = 0;
 
 	    StringCchLengthW(pInfo->UserName, MAX_USERNAME_LENGTH, &szlen);
-	    WideCharToMultiByte(CP_UTF8, 0, pInfo->UserName, szlen,
+	    WideCharToMultiByte(CP_UTF8, 0, pInfo->UserName, (int)szlen,
 				 username, sizeof(username), NULL, NULL);
 
 	    StringCchLengthW(pInfo->Domain, MAX_DOMAIN_LENGTH, &szlen);
-	    WideCharToMultiByte(CP_UTF8, 0, pInfo->Domain, szlen,
+	    WideCharToMultiByte(CP_UTF8, 0, pInfo->Domain, (int)szlen,
 				 domain, sizeof(domain), NULL, NULL);
 
 	    GetDomainLogonOptions(NULL, username, domain, &opt);
@@ -1426,11 +1426,11 @@ VOID AFS_Logon_Event( PWLX_NOTIFICATION_INFO pInfo )
 	DebugEvent0("AFS_Logon_Event - pInfo UserName and Domain");
 
         StringCchLengthW(pInfo->UserName, MAX_USERNAME_LENGTH, &szlen);
-        WideCharToMultiByte(CP_UTF8, 0, pInfo->UserName, szlen,
+        WideCharToMultiByte(CP_UTF8, 0, pInfo->UserName, (int)szlen,
                             username, sizeof(username), NULL, NULL);
 
         StringCchLengthW(pInfo->Domain, MAX_DOMAIN_LENGTH, &szlen);
-        WideCharToMultiByte(CP_UTF8, 0, pInfo->Domain, szlen,
+        WideCharToMultiByte(CP_UTF8, 0, pInfo->Domain, (int)szlen,
                             domain, sizeof(domain), NULL, NULL);
 
 	DebugEvent0("AFS_Logon_Event - Calling GetDomainLogonOptions");
